Remove value check in 'disabled' class handling for jQuery BS-Select
The update removes the check for options without any value in the handling of 'disabled' class within the jQuery Bootstrap Select component. This change prevents the immediate return for options without values and may affect how we handle these options in subsequent code lines. Code readability and performance are also improved due to adjustments in formatting.