Support theme attribute in jquery.bs-select.js dropdown
This commit adds support for a 'data-bs-theme'
attribute within the jQuery bs-select plugin.
Now when a dropdown menu is created,
if the select element has a parent 'data-bs-theme'
attribute, it will be copied to the newly created dropdown container.
This allows for more flexible theming of the generated dropdown menus.