Add 'onKeyDown' function and related event
Added a new 'onKeyDown' function and 'keydown.bs.select' event in the code. This function is triggered when a key that is not a dropdown command is pressed. This feature enhances the user interaction by providing response to non-dropdown keys. Furthermore, a few minor formatting and code cleanup tasks have been performed.